Removed platform dependencies
Removed all platform-specific and legacy dependencies like Windows Forms, WPF, `System.Drawing` or `DevIL`. The base library `GlobalLib` is now free of any dependencies (except for `Microsoft.CSharp` which is required for dynamic memory casts) and will be the core API for the project and therefore will be renamed in the future. Additional functionality like exporting textures/geometry will be implemented in separate projects/assemblies to keep everything as modular as possible.