Moved to C# 9
Although the new language features of C# 12 are great, we're moving back to C# 9 to ensure maximum compatibility. Although the language version only affects the compile step and not the compatibility of the library build, it enables developers to use the library's source code instead of the compiled assemblies for a greater range of projects.