Cleaned codebase
Tidied up the whole codebase so it adheres to (modern) coding standards and does not produce a huge amount of code inspection warnings. The next goal is to make everything `private` (or `internal`/`protected`) that does not need to be `public` and move further away from reflection (wherever reasonable) in favour of polymorphism.